refactor: Move CardListComponent tests to tests directory
- Move all 41 CardListComponent unit tests to tests/card_list_component.rs
- Tests now run as integration tests instead of inline unit tests
- Maintains same test file structure mirroring source code organization
- All 41 integration tests passing
- Full test suite: 75 total tests passing

Benefits:
- Cleaner separation of test code from source code
- Easier to maintain and organize tests
- Integration tests provide better isolation
- Source files now focus on implementation only
